requirements.txt
colorama==0.4.5
pretty-errors==1.2.25
python-vlc==3.0.16120
